LinkedList Changes
Added an enforced nodeType variable to LinkedList.  Whenever the list is
empty, nodeType is equal to nil.  Upon the first addition to the list,
nodeType is set to the object's type.  Every following addition must be
of the same type as the first.

Also, there was an addition to the LLNode class.  I added a metamethod
for table access.  If the 'value' data member is a table, you can
directly call the function from the node.

EX) If LLNode's are holding UIObjects we can do this for drawing:

for node in lList:Iterate() do
node:Draw() -- assuming each value of node has a 'Draw' function
end