UIObject Updates
UIComponent is no more.  Instead, the abstract methods are placed in
UIBase.  There are two folders in core/gui now for loading purposes:
core/gui/base and core/gui/components.  Base is for uiBase.lua and
uiDraggable.lua.  Anything that will be extended on the base UIObjects
needs to be placed here so there are no errors trying to extend a class
that hasn't been loaded yet.